+++ libs/platform/unit/platform_unit_test.cpp
@@ -569,7 +569,7 @@ TEST_CASE("access_check", "[platform]")
     _test_helper test_helper;
     ebpf_security_generic_mapping_ptr sd_ptr;
     ebpf_security_descriptor_t* sd = NULL;
-    DWORD sd_size = 0;
+    unsigned long sd_size = 0;
     ebpf_security_generic_mapping_t generic_mapping{1, 1, 1};
     auto allow_sddl = L"O:COG:BUD:(A;;FA;;;WD)";
     auto deny_sddl = L"O:COG:BUD:(D;;FA;;;WD)";
